backend/pyproject.toml
[project]
name = "proofread-backend"
version = "0.1.0"
description = "ProofRead backend — TTB label compliance verification"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0.27",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pyyaml>=6.0",
    "rapidfuzz>=3.6",
    "python-multipart>=0.0.9",
    "boto3>=1.34",
    "Pillow>=10.2",
    "anthropic>=0.40",
    "jinja2>=3.1",
    "numpy>=1.26",
    "scipy>=1.11",
]

[project.optional-dependencies]
google-vision = ["google-cloud-vision>=3.7"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "ruff>=0.3",
    "mypy>=1.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "ASYNC"]
# UP042: keeping `str, Enum` mixin (not StrEnum) — call sites use explicit
#   `.value` access and changing `__str__` semantics could affect serialization.
ignore = ["UP042"]

[tool.ruff.lint.per-file-ignores]
# B008: FastAPI's `Depends(...)`, `File(...)`, `Form(...)` are idiomatic
#   in route handler defaults; the framework expects them at call-construction
#   time. Suppressed only in API routers.
"app/api/*.py" = ["B008"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
